+++ 2022/advent_of_code/src/days/day8.rs
@@ -17,19 +17,19 @@ pub fn day8(input_lines: Vec<&str>) {
     // Reveal left to right and right to left
     for y in 0..height {
         // Left to right
-        flag_visible(&mut trees, width, height, Point::new(0, y as i32), Point::new(1, 0));
+        flag_visible(&mut trees, width, height, Point::new(0, y as i64), Point::new(1, 0));
 
         // Right to left
-        flag_visible(&mut trees, width, height, Point::new(width as i32- 1, y as i32), Point::new(-1, 0));
+        flag_visible(&mut trees, width, height, Point::new(width as i64- 1, y as i64), Point::new(-1, 0));
     }
 
     // Reveal up to down and down to up
     for x in 0..width {
         // Up to down
-        flag_visible(&mut trees, width, height, Point::new(x as i32, 0), Point::new(0, 1));
+        flag_visible(&mut trees, width, height, Point::new(x as i64, 0), Point::new(0, 1));
 
         // Down to up
-        flag_visible(&mut trees, width, height, Point::new(x as i32, height as i32 -1), Point::new(0, -1));
+        flag_visible(&mut trees, width, height, Point::new(x as i64, height as i64 -1), Point::new(0, -1));
     }
 
     let mut visible_count = 0;
@@ -48,7 +48,7 @@ pub fn day8(input_lines: Vec<&str>) {
 
     for y in 0..height {
         for x in 0..width {
-            let point = Point::new(x as i32, y as i32);
+            let point = Point::new(x as i64, y as i64);
             let score = 
                 // Left to right
                 get_scenic_score_in_direction(&trees, width, height, point, Point::new(1, 0)) *
@@ -75,7 +75,7 @@ fn flag_visible(trees: &mut Vec<Vec<Tree>>, width: usize, height: usize, start: 
     let mut point = start;
     while 
         point.x >= 0 && point.y >= 0 &&
-        point.x < width as i32 && point.y < height as i32 {
+        point.x < width as i64 && point.y < height as i64 {
 
         let tree = &mut trees[point.x as usize][point.y as usize];
         if tree.height > max_height {
@@ -96,7 +96,7 @@ fn get_scenic_score_in_direction(trees: &Vec<Vec<Tree>>, width: usize, height: u
 
     while 
         point.x >= 0 && point.y >= 0 &&
-        point.x < width as i32 && point.y < height as i32 {
+        point.x < width as i64 && point.y < height as i64 {
             
         score += 1;
